t/: don't use .git/COMMIT_EDITMSG
Some tests in t3507-cherry-pick-conflict.sh and t7502-commit.sh rely
on 'git commit' leaving a stale .git/COMMIT_EDITMSG file behind even
on success, and check the file's content to verify that the commit
message template was set up correctly.

Use a special GIT_EDITOR that makes a copy of COMMIT_EDITMSG and
modify those checks to work on that copy, so the next patch can teach
'git commit' to not leave a stale COMMIT_EDITMSG file behind.

Signed-off-by: SZEDER Gábor <szeder.dev@gmail.com>