line-log: get rid of the diffqueues array in process_ranges_ordinary_commit()
In process_ranges_merge_commit() we used to have two loops: one to
compute diffs between a merge commit and each of its parents and store
them in an array of 'struct diff_queue_struct', and the other to
process all diffs in that array.  Since the refactoring done in an
earlier patch in this series there is only one loop left, which both
computes and processes a diff on each iteration, so that array is not
necessary anymore.

Remove that array and use a single loop-local instance of 'struct
diff_queue_struct' instead, thereby simplifying code and avoiding a
memory allocation for each processed merge commit.

With this change the static free_diffqueues() helper in 'line-log.c'
becomes callerless, remove it.

Signed-off-by: SZEDER Gábor <szeder.dev@gmail.com>