test-lib-functions: use BUG() instead of 'error "bug in the test script"'
The BUG() function of our test framework is essentially a shorthand
for 'error "bug in the test script: ..."' and a useful redirection to
the test script's original standard error [1], but there are two
places where we still use that 'error' command without redirection.

Use BUG() in these two cases as well.

[1] 165293af3c (tests: send "bug in the test script" errors to the
    script's stderr, 2018-11-19)

Signed-off-by: SZEDER Gábor <szeder.dev@gmail.com>